refactor: 💡 split code into two hooks
